+++ sentry-samples/sentry-samples-android/src/main/java/io/sentry/samples/android/MainActivity.java
@@ -8,7 +8,6 @@ import io.sentry.ISpan;
 import io.sentry.MeasurementUnit;
 import io.sentry.Sentry;
 import io.sentry.UserFeedback;
-import io.sentry.android.core.SentryAndroid;
 import io.sentry.instrumentation.file.SentryFileOutputStream;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.User;
@@ -198,19 +197,7 @@ public class MainActivity extends AppCompatActivity {
         });
 
     setContentView(binding.getRoot());
-
-    // Let's say the activity is fully drawn after 1 second
-    new Thread(
-            () -> {
-              try {
-                Thread.sleep(1000);
-              } catch (InterruptedException e) {
-                e.printStackTrace();
-              }
-              SentryAndroid.reportFullyDrawn(this);
-              reportFullyDrawn();
-            })
-        .start();
+    Sentry.reportFullDisplayed();
   }
 
   @Override