+++ sentry-samples/sentry-samples-android/src/main/java/io/sentry/samples/android/ProfilingActivity.kt
@@ -69,6 +69,7 @@ class ProfilingActivity : AppCompatActivity() {
             }.start()
         }
         setContentView(binding.root)
+        Sentry.reportFullDisplayed()
     }
 
     private fun finishTransactionAndPrintResults(t: ITransaction) {