+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -405,4 +405,14 @@ class SentryOptionsTest {
         options.transactionPerformanceCollector = performanceCollector
         assertEquals(performanceCollector, options.transactionPerformanceCollector)
     }
+
+    @Test
+    fun `when options are initialized, TimeToFullDisplayTracing is false`() {
+        assertFalse(SentryOptions().isEnableTimeToFullDisplayTracing)
+    }
+
+    @Test
+    fun `when options are initialized, FullyDrawnReporter is set`() {
+        assertEquals(FullyDisplayedReporter.getInstance(), SentryOptions().fullyDrawnReporter)
+    }
 }